Moved SearchAndChoose into a Fragment; Fixed Done button on LoginFragment
