+++ app/src/main/java/nyc/c4q/ashiquechowdhury/auxx/LoginFragment.java
@@ -4,11 +4,14 @@ import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v7.widget.CardView;
+import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.view.inputmethod.EditorInfo;
 import android.widget.EditText;
 import android.widget.ImageView;
+import android.widget.TextView;
 import android.widget.Toast;
 
 import com.bumptech.glide.Glide;
@@ -16,6 +19,8 @@ import com.bumptech.glide.Glide;
 public class LoginFragment extends Fragment {
     private EditText usernameEditText;
     private EditText passwordEditText;
+    String enteredUsername;
+    String enteredPassword;
 
     @Nullable
     @Override
@@ -32,15 +37,26 @@ public class LoginFragment extends Fragment {
         CardView googleSignInButton = (CardView) view.findViewById(R.id.google_signin_button);
         usernameEditText = (EditText) view.findViewById(R.id.username_login_edittext);
         passwordEditText = (EditText) view.findViewById(R.id.password_login_edittext);
+
+        passwordEditText.setOnEditorActionListener(new EditText.OnEditorActionListener() {
+            @Override
+            public boolean onEditorAction(TextView textView, int actionId, KeyEvent keyEvent) {
+                if (actionId == EditorInfo.IME_ACTION_DONE) {
+                    enteredUsername = getUsernameText();
+                    enteredPassword = getPasswordText();
+                    handleLogin(enteredUsername, enteredPassword);
+                    return true;
+                }
+                return false;
+            }
+        });
+
         auxSignInButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                String enteredUsername = getUsernameText();
-                String enteredPassword = getPasswordText();
-
-//                handleLogin(enteredUsername, enteredPassword);
-                getActivity().getSupportFragmentManager().beginTransaction()
-                        .replace(R.id.main_container, new JoinRoomFragment()).commit();
+                enteredUsername = getUsernameText();
+                enteredPassword = getPasswordText();
+                handleLogin(enteredUsername, enteredPassword);
             }
         });
 
@@ -67,7 +83,8 @@ public class LoginFragment extends Fragment {
         } else if (enteredPassword.isEmpty()) {
             Toast.makeText(getActivity(), "Enter a password", Toast.LENGTH_SHORT).show();
         } else {
-
+            getActivity().getSupportFragmentManager().beginTransaction()
+                    .replace(R.id.main_container, new JoinRoomFragment()).commit();
         }
     }
 }
