+++ app/src/main/java/nyc/c4q/ashiquechowdhury/auxx/SlidePanelActivity.java
@@ -1,0 +1,18 @@
+package nyc.c4q.ashiquechowdhury.auxx;
+
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.support.v7.app.AppCompatActivity;
+
+/**
+ * Created by SACC on 3/5/17.
+ */
+
+public class SlidePanelActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_searchandchoose_container);
+    }
+}
