+++ app/src/main/res/layout/activity_searchandchoose_innerframe.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/searchandchoose_innerframe"
+    android:background="#FF00"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+</FrameLayout>
